refactor: broke down project service into smaller pieces
service/project.py became too large to comprehend with
the direct packages references from git gets ignored,
hence broke down it into smaller pieces.

GH-94
