+++ src/ossiq/commands/html.py
@@ -8,7 +8,7 @@ from typing import Literal
 import typer
 
 from ossiq.domain.common import Command, UserInterfaceType
-from ossiq.service import project
+from ossiq.service.project.scan import scan
 from ossiq.settings import Settings
 from ossiq.sources import project_sources
 from ossiq.ui.registry import get_renderer
@@ -58,7 +58,7 @@ def command_html(ctx: typer.Context, options: CommandHtmlOptions) -> None:
     )
 
     with show_scan_progress(settings) as on_step:
-        project_scan = project.scan(sources, on_step=on_step)
+        project_scan = scan(sources, on_step=on_step)
 
     renderer = get_renderer(command=Command.HTML, user_interface_type=UserInterfaceType.HTML, settings=settings)
