+++ src/ossiq/service/update.py
@@ -6,7 +6,7 @@ import dataclasses
 from dataclasses import dataclass, field
 
 from ossiq.domain.common import ConstraintType
-from ossiq.service.project import ScanRecord, ScanResult
+from ossiq.service.project.models import ScanRecord, ScanResult
 from ossiq.service.update_impact import TransitiveImpact
 from ossiq.solver.reason import RecommendationReason
